requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dictdiffer==0.9.0
distlib==0.3.8
Events==0.5
filelock==3.13.1
humanize==4.9.0
idna==3.6
loguru==0.6.0
pipenv==2023.12.1
platformdirs==4.2.0
pydantic==1.9.0
pydavinci-18 @ git+https://github.com/in03/pydavinci.git@51fc6708b6295fe02d093094c0f171614ae7c8aa
PyYAML==6.0.1
-e git+ssh://git@github.com/seb26/renderwatch.git@616e1fb8396632564bd2f83ecd86b3d9a8f727e6#egg=renderwatch
requests==2.31.0
typing_extensions==4.1.1
urllib3==2.1.0
virtualenv==20.25.1
yamale==5.0.0
